Add check on date to ensure srs works as expected
When I clicked on a card that had an availableDate that was days prior
to the present day, when I gave a quality response of 4 or more the date
was getting updated to the next day, which was in the past, or a day
before the present, so the card was still in the queue.
The desired effect would be that the card would be due the next day or
in the future relative to the present day.

So now if the date when the card is next due is before now, the date
will be set to Date.now()

This is slightly redundant because the app would never allow you to do
a PUT request on a card that was due on the future in its current state.
However, for development purposes, I might want to alter the data
through a PUT request without setting the date to Date.now()

Sorry for writing so much...