Add feature to upload csv files to make decks
I had a lot of trouble with this, and I installed a bunch of node
packages to make this work, but in the end it I only really needed
the csv package. That is, if I'm not mistaken. The documentation for
the csv package is trash though that should not have been so difficult.
Okay it's not trash, I just really struggled to understand its purpose
and how to use it. It could really use some improvements though.

Unless an issue arises this feature appears to be pretty much complete,
except for the UI